app/requirements.txt
exponent-server-sdk
fastapi
pydantic
pydantic_core
pytest
python-dotenv
supabase
uvicorn
